Removed com.sun,mirror dependency, and rename jsr269 packages to original.
