--- slim3-gen-jsr269/src/main/java/org/slim3/jsr269/util/DeclarationUtil.java
@@ -1,91 +1,0 @@
-/*
- * Copyright 2004-2010 the Seasar Foundation and the Others.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
- * either express or implied. See the License for the specific language
- * governing permissions and limitations under the License.
- */
-package org.slim3.jsr269.util;
-
-import javax.lang.model.element.Element;
-import javax.lang.model.element.ElementKind;
-import javax.lang.model.element.ExecutableElement;
-import javax.lang.model.element.TypeElement;
-
-import org.slim3.gen.processor.UnknownDeclarationException;
-
-import com.sun.mirror.apt.AnnotationProcessorEnvironment;
-import com.sun.mirror.declaration.AnnotationMirror;
-import com.sun.mirror.declaration.AnnotationTypeDeclaration;
-import com.sun.mirror.declaration.Declaration;
-
-/**
- * A utility class for operationg declarations.
- * 
- * @author taedium
- * @since 1.0.0
- * 
- */
-@SuppressWarnings("deprecation")
-public final class DeclarationUtil {
-
-    /**
-     * Returns {@code AnnotationMirror} if a declaration is annotated with a
-     * specified annotation and {@code null} otherwise.
-     * 
-     * @param env
-     *            the environment
-     * @param declaration
-     *            the declaration object to be checked.
-     * @param annotation
-     *            the fully qualified name of an annotation.
-     * @return {@code AnnotationMirror} if an declaration is annotated with a
-     *         specified annotation and {@code null} otherwise.
-     */
-    public static AnnotationMirror getAnnotationMirror(
-            AnnotationProcessorEnvironment env, Declaration declaration,
-            final String annotation) {
-        if (declaration == null) {
-            throw new NullPointerException("The declaration parameter is null.");
-        }
-        for (AnnotationMirror mirror : declaration.getAnnotationMirrors()) {
-            AnnotationTypeDeclaration annotationTypeDeclaration =
-                mirror.getAnnotationType().getDeclaration();
-            if (annotationTypeDeclaration == null) {
-                throw new UnknownDeclarationException(env, declaration, mirror);
-            }
-            if (annotationTypeDeclaration.getQualifiedName().equals(annotation)) {
-                return mirror;
-            }
-        }
-        return null;
-    }
-
-    /**
-     * Returns {@code true} if the class declaration has a public default
-     * constructor.
-     * 
-     * @param classElement
-     *            the class declaration
-     * @return if the class declaration has a public default constructor.
-     */
-    public static boolean hasPublicDefaultConstructor(TypeElement classElement) {
-        for (Element el : classElement.getEnclosedElements()) {
-            if (el.getKind() == ElementKind.CONSTRUCTOR) {
-                ExecutableElement ex = (ExecutableElement) el;
-                if (ex.getParameters().isEmpty()) {
-                    return true;
-                }
-            }
-        }
-        return false;
-    }
-}
