--- slim3-gen-jsr269/src/main/resources/META-INF/services/com.sun.mirror.apt.AnnotationProcessorFactory
@@ -1,1 +1,0 @@
-org.slim3.gen.processor.ModelProcessorFactory
